Move some files aroudn
