+++ packages/@aws-cdk/toolkit-lib/lib/actions/index.ts
@@ -9,3 +9,4 @@ export * from './refactor';
 export * from './rollback';
 export * from './synth';
 export * from './watch';
+export * from './diagnose';
