--- packages/@aws-cdk/toolkit-lib/lib/api/diagnose/index.ts
@@ -1,1 +1,0 @@
-export * from './types';
