+++ packages/@aws-cdk/toolkit-lib/lib/api/index.ts
@@ -23,5 +23,5 @@ export * from './work-graph';
 export * from './tree';
 export * from './tags';
 export * from './drift';
-export * from './diagnose';
+export * from '../actions/diagnose';
 export * from './source-tracing';
